+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `read: bool` option on all_steps to force read from storage location instead
 of using cached steps.
 - Simplified "Getting Started" section in the docs.
+- CheckpointManager creates the top level directory if it does not yet exist.
 
 ### Changed
 - Removed some unused test_utils methods for filtering empty nodes.